backend/spdb/requirements.txt
asgiref==3.2.7
Django==3.0.6
djangorestframework==3.11.0
django-cors-headers==3.3.0
flake8==3.8.2
importlib-metadata==1.6.0
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
ortools
numpy